+++ src/main/scala/com/guizmaii/scalajwt/utils/ProvidedValidations.scala
@@ -14,7 +14,7 @@ object ProvidedValidations {
     * The DefaultJWTClaimsVerifier will check the token expiration vut only if `exp` claim is present.
     * We could need to require its presence.
     */
-  val requireExpirationClaim: (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
+  final val requireExpirationClaim: (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
     (jwtClainSet: JWTClaimsSet, _: SecurityContext) => {
       if (jwtClainSet.getExpirationTime == null)
         Some(MissingExpirationClaim)
@@ -25,7 +25,7 @@ object ProvidedValidations {
   /**
     * Will ensure that the `token_use` claim is equal to the passed String value.
     */
-  val requireTokenUseClaim: (String) => (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
+  final val requireTokenUseClaim: (String) => (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
     (requiredTokenUseValue: String) =>
       (jwtClainSet: JWTClaimsSet, _: SecurityContext) => {
         val tokenUse: String = jwtClainSet.getStringClaim("token_use")
@@ -38,7 +38,7 @@ object ProvidedValidations {
   /**
     * Will ensure that the `iss` claim contains the passed String value.
     */
-  val requiredIssuerClaim: (String) => (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
+  final val requiredIssuerClaim: (String) => (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
     (requiredIssuerValue: String) =>
       (jwtClainSet: JWTClaimsSet, _: SecurityContext) => {
         val iss: String = jwtClainSet.getIssuer
@@ -51,7 +51,7 @@ object ProvidedValidations {
   /**
     * Will ensure that the `sub` claim is present.
     */
-  val requiredNonEmptySubject: (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
+  final val requiredNonEmptySubject: (JWTClaimsSet, SecurityContext) => Option[BadJWTException] =
     (jwtClainSet: JWTClaimsSet, _: SecurityContext) => {
       val userId: String = jwtClainSet.getSubject
       if (userId == null || userId.isEmpty)